refactor: move project service into contentlayer lib
